requirements/analysis-requirements.txt
dash-cytoscape==0.3.0
dash==1.21.0
matplotlib==3.4.2
networkx==2.5
pandas==1.1.3
plotly==4.14.3
psycopg2-binary==2.8.6
scipy==1.7.3
wait-for-it==2.2.1
